Delete unused class and rename class
