+++ GeoBot/GeoBot/Helpers/Token.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace GeoBot.Helpers
+{
+    public class Token
+    {
+        public string token { get; set; }
+    }
+
+}
